backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
openai==1.3.0
runware>=0.4.30
pillow>=10.0.0
python-dotenv>=1.0.0
aiofiles==23.2.1
aiohttp>=3.9.0
